% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ld}
\alias{ld}
\title{Lexical decision data}
\format{An object of class \code{data.frame} with 25401 rows and 6 columns.}
\source{
Keuleers, E., Lacey, P., Rastle, K., & Brysbaert, M. (2012). The British Lexicon 
Project: Lexical decision data for 28,730 monosyllabic and disyllabic English words. 
Behavior Research Methods, 44(1), 287-304.
}
\usage{
ld
}
\description{
Lexical decision data from the British Lexicon Project (Keuleers et al., 2012)
}
\keyword{datasets}
