% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nam}
\alias{nam}
\title{Word naming data}
\format{An object of class \code{data.frame} with 37107 rows and 6 columns.}
\source{
Balota, D. A., Yap, M. J., Hutchison, K. A., Cortese, M. J., Kessler, B., Loftis, B., ... 
& Treiman, R. (2007). The English lexicon project. Behavior Research Methods, 39(3), 445-459.
}
\usage{
nam
}
\description{
Word naming data from the English Lexicon Project (Balota et al., 2007)
}
\keyword{datasets}
