% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam_ld}
\alias{pam_ld}
\title{Lexical decision PAM}
\format{An object of class \code{gam} (inherits from \code{glm}, \code{lm}) of length 51.}
\source{
Keuleers, E., Lacey, P., Rastle, K., & Brysbaert, M. (2012). The British Lexicon 
Project: Lexical decision data for 28,730 monosyllabic and disyllabic English words. 
Behavior Research Methods, 44(1), 287-304.
}
\usage{
pam_ld
}
\description{
PAM fit to the lexical decision data in the British Lexicon Project (Keuleers et al., 2012)
}
\keyword{datasets}
