% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrier.R
\name{plotBrier}
\alias{plotBrier}
\title{Plot Brier}
\usage{
plotBrier(
  pec,
  type = "BS",
  main = NA,
  xlab = NA,
  ylab = NA,
  ylim = NA,
  colors = c("#B33333", "#3333B3", "#33B333", "#B333B3", "#B37733"),
  lwd = 2,
  legend.names = NA,
  y.intersp = 1,
  ...
)
}
\arguments{
\item{pec}{A pec object, as created by pec()}

\item{type}{"BS" for Brier score (default), "IBS" for integrated Brier score}
}
\description{
Plot the Brier score
}
\examples{
# Remove outliers
predictors = c("logFrequency", "Length", "logOLD20", "SND20") 
ld = removeOutliers(ld, predictors)
ld = na.omit(ld)

# Prepare data in exponential data format
ld$status = 1
cut_points = as.numeric(quantile(ld$RT[which(ld$RT <= 1085 & 
               ld$RT >= 500)],seq(0, 1, by = 0.02)))
ped = split_data(Surv(RT, status)~., data = ld, id = "id",
                   cut = cut_points)

# Run PAM (warning: computationally heavy)
pam_ld = pamm(ped_status ~ s(tend) + 
             s(logFrequency) + ti(tend, logFrequency) + 
             s(Length) + ti(tend, Length) + 
             s(logOLD20) + ti(tend, logOLD20) + 
             s(SND20) + ti(tend, SND20),
             data = ped)

# Generate prediction error curves
pec = pec(list(pam = pam_ld),
  Surv(RT, status) ~ 1,
  data = na.omit(ld),
  times = seq(510, 1080, length.out = 100),
  start = 510,
  exact = FALSE)

# Plot Brier score
plotBrier(pec)

# Plot integrated Brier score
plotBrier(pec, type = "IBS")

}
