% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQGAMs.R
\name{plotQGAMs}
\alias{plotQGAMs}
\title{Plot quantiles}
\usage{
plotQGAMs(
  models,
  predictor,
  data,
  cols = c("#000080", "#1A1A9A", "#3333B3", "#4D4DCD", "#6666E6"),
  se = 2,
  xlab = NA,
  ylab = NA,
  main = NA,
  ylim = NA,
  ...
)
}
\arguments{
\item{models}{A list of QGAM models as generated by the mqgam() function in the qgam package.}

\item{predictor}{The predictor to be plotted. This predictor needs to be present in the fitted 
models, as well as in data.}

\item{data}{The data the QGAM models were fit to. Needs to include the response variable, 
as well as all predictors in these models.}

\item{cols}{A vector of colors. The lines corresponding to the quantiles will be plotted in 
these colors.}

\item{se}{The number of standard errors for the confidence intervals. Default: 2 (i.e., 95\% confidence intervals)}
}
\description{
Plot the results of a series of QGAM models (Fasiolo et al., 2017)
}
\examples{
# Remove outliers from the ld data set, which contains lexical 
# decision latencies from the British Lexicon Project (Keuleers 
# et al, 2012)
predictors = c("RT", "logFrequency", "Length", "logOLD20", "SND20")
ld = removeOutliers(ld, predictors)
ld = na.omit(ld)

# Tune learning rate for median
tune = tuneLearnFast(RT ~ s(logFrequency) + s(Length) + 
                          s(logOLD20) + s(SND20),
                     data = ld, qu = 0.5)
sigpar = tune$lsig

# Define quantiles
quants = c(0.10,0.25,0.50,0.75,0.90)

# Run qgam models
qgams = mqgam(RT ~ s(logFrequency) + s(Length) + s(logOLD20) + 
                   s(SND20),
                   data = ld, qu = quants, lsig = sigpar)

# Plot effect of frequency at quantiles
plotQGAMs(qgams, "logFrequency", ld)

}
\references{
Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast
calibrated additive quantile regression. URL: 
https://arxiv.org/abs/1707.03307.

Keuleers, E., Lacey, P., Rastle, K., & Brysbaert, M. (2012). 
The British Lexicon Project: Lexical decision data for 28,730 
monosyllabic and disyllabic English words. Behavior Research 
Methods, 44(1), 287-304.
}
