% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pn}
\alias{pn}
\title{Picture naming data}
\format{An object of class \code{data.frame} with 484 rows and 6 columns.}
\source{
Székely, A., D’amico, S., Devescovi, A., Federmeier, K., Herron, D., Iyer, G., ... & Bates, E. (2003). 
Timed picture naming: Extended norms and validation against previous studies. Behavior Research Methods, 
Instruments, & Computers, 35(4), 621-633.
}
\usage{
pn
}
\description{
Picture naming data  (Balota et al., 2007)
}
\keyword{datasets}
