% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutliers.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove outliers}
\usage{
removeOutliers(data, predictors, sd = 3)
}
\arguments{
\item{data}{A data frame.}

\item{predictors}{A vector of the column names in data corresponding to the predictors that outliers 
should be removed for
sd The number of standard deviations. Predictor values further than sd standard deviations 
from the predictor mean are removed. Default: 3.}

\item{model}{A PAM model.}
}
\description{
Remove outliers from a data frame
}
\examples{
# Load data for the lexical decision latencies in the British Lexicon Project (Keuleers et al, 2012)
data(ld)

# Remove outliers
ld = removeOutliers.fnc(ld, c("logFrequency","Length","logOLD20","SND20"))
ld = na.omit(ld)

}
