% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pn}
\alias{pn}
\title{Picture naming data}
\format{

}
\source{
Székely, A., D’amico, S., Devescovi, A., Federmeier, K., Herron, D., Iyer, G., ... & Bates, E. (2003). 
Timed picture naming in seven languages. Bates, E., D'Amico, S., Jacobsen, T., Szekely, A., 
Andonova, E., et al. (2003). Psychonomic Bulletin & Review, 10(2), 344-380.
}
\usage{
pn
}
\description{
Picture naming data  (Bates et al., 2003)
}
\keyword{datasets}
